TeraManager Version 0.5.1
November 26th, 2012

--------------------------------------------------------------------------------------------
- New: color map changes are propagated when switching between different resolutions in the
       multiresolution mode.
--------------------------------------------------------------------------------------------

+++ hackathon/teramanager/core/ImageManager/ProgressBar.cpp
@@ -41,8 +41,8 @@ void ProgressBar::update(float new_progress_value, const char* new_progress_info
 
 	if(new_progress_value!=0)
 	{
-		minutes_remaining = (proctime + TIME(0))*(100.0-progress_value)/(progress_value*60.0);
-		seconds_remaining = (proctime + TIME(0))*(100.0-progress_value)/(progress_value);
+		minutes_remaining = (int) ((proctime + TIME(0))*(100.0-progress_value)/(progress_value*60.0));
+		seconds_remaining = (int) ((proctime + TIME(0))*(100.0-progress_value)/(progress_value));
 	}
 }
 
@@ -65,7 +65,7 @@ void ProgressBar::show()
 	printf("%d%%\t",(int)(progress_value));
 	for(int i=1; i<=progress_value; i++)
 		printf("Û");
-	for(int i=progress_value; i<100; i++)
+	for(int i=(int)progress_value; i<100; i++)
 		printf(":");
 	printf("\n\n");
 }
